ijjgfr

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -6,7 +6,7 @@
 */
 void print_number(int n)
 {
-	unsigned int k, s, i, l = 1, r = 0;
+	unsigned long int k, s, i, l = 1, r = 0;
 
 	if (n >= 0)
 	{
